+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add support for `stopped` and `cancelled` jobs.
 ### Bug Fixes
 - Fixed a bug that did not allow filtering jobs by status value.
+- Fixed issue with deleting a stopped jobs.
 
 ## [0.4.0](https://github.com/dldevinc/paper-rq/tree/v0.4.0) - 2022-01-13
 ### ⚠ BREAKING CHANGES